requirements.txt
selenium
pygsheets
google-cloud-translate==2.0.1
